Modified   assets/scripts/BuiltinBox.hlsl Modified   assets/scripts/BuiltinBox.material Modified   assets/scripts/BuiltinBox.tsc Modified   assets/scripts/BuiltinBox.vs Modified   source/step09/Core/Include/Parser/T3DScriptParser.h Modified   source/step09/Core/Include/Render/T3DRenderer.h Modified   source/step09/Core/Include/Render/T3DRenderer.inl Modified   source/step09/Core/Include/Tiny3D.h Modified   source/step09/Plugins/Renderer/Direct3D11/Include/T3DD3D11Renderer.h Modified   source/step09/Plugins/Renderer/Direct3D11/Source/T3DD3D11Renderer.cpp Modified   source/step09/Plugins/Renderer/Reference3D/Include/T3DR3DRenderer.h Modified   source/step09/Plugins/Renderer/Reference3D/Source/T3DR3DRenderer.cpp Modified   source/step09/Tools/ScriptCompiler/Source/T3DScriptCompiler.cpp

+++ source/step09/Core/Include/Parser/T3DScriptParser.h
@@ -36,7 +36,7 @@ namespace Tiny3D
         {
             char        magic[4];   /**< 文件魔数，用于标识文件 */
             uint32_t    version;    /**< 文件版本号 */
-            uint32_t    filesize;   /**< 文件大小 */
+            size_t      filesize;   /**< 文件大小 */
         };
 
     public:
